+++ OSX/MacGUI/C64Proxy.mm
@@ -165,18 +165,19 @@ struct CartridgeWrapper { Cartridge *cartridge; };
 
 - (void *) screenBuffer { return wrapper->vic->screenBuffer(); }
 
-- (NSColor *) color:(int)nr
+- (NSColor *) color:(NSInteger)nr
 {
     assert (0 <= nr && nr < 16);
     
-    uint32_t color = wrapper->vic->getColor(nr);
+    uint32_t color = wrapper->vic->getColor((unsigned)nr);
     uint8_t r = color & 0xFF;
     uint8_t g = (color >> 8) & 0xFF;
     uint8_t b = (color >> 16) & 0xFF;
     
 	return [NSColor colorWithCalibratedRed:(float)r/255.0 green:(float)g/255.0 blue:(float)b/255.0 alpha:1.0];
 }
-- (void) setColor:(unsigned)nr rgba:(int)rgba { wrapper->vic->setColor(nr, rgba); }
+- (void) setColor:(NSInteger)nr rgba:(uint32_t)rgba {
+    wrapper->vic->setColor((unsigned)nr, rgba); }
 
 - (uint16_t) memoryBankAddr { return wrapper->vic->getMemoryBankAddr(); }
 - (void) setMemoryBankAddr:(uint16_t)addr { wrapper->vic->setMemoryBankAddr(addr); }
